--- app/javascript/scavenger_hunt/scenes/Questions/QuestionEight/index.js
@@ -1,1 +1,0 @@
-export { QuestionEight } from './QuestionEight';
