+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionEleven/QuestionEleven.scss
