+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionEleven/index.js
@@ -1,0 +1,1 @@
+export { QuestionEleven } from './QuestionEleven';
