+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionNine/QuestionNine.scss
@@ -8,11 +8,11 @@
   margin-top: 20px;
 }
 
-.emo {
- width: 100%;
+.table {
+    width: 80%;
+    height: 300px;
 }
 
-.pictionary {
-  margin-top: 10px;
-  margin-bottom: 10px;
+.seq {
+  font-size: 12px;;
 }
