+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionTen/QuestionTen.scss
@@ -8,11 +8,11 @@
   margin-top: 20px;
 }
 
-.table {
-    width: 80%;
-    height: 300px;
+.emo {
+ width: 100%;
 }
 
-.seq {
-  font-size: 12px;;
+.pictionary {
+  margin-top: 10px;
+  margin-bottom: 10px;
 }
