+++ app/javascript/scavenger_hunt/scenes/Questions/index.js
@@ -5,6 +5,6 @@ export { QuestionFour } from './QuestionFour';
 export { QuestionFive } from './QuestionFive';
 export { QuestionSix } from './QuestionSix';
 export { QuestionSeven} from './QuestionSeven';
-export { QuestionEight } from './QuestionEight';
 export { QuestionNine } from './QuestionNine';
 export { QuestionTen } from './QuestionTen';
+export { QuestionEleven } from './QuestionEleven';
